session issue

+++ src/main/resources/static/js/welcome.js
@@ -1,8 +1,16 @@
-let username = window.prompt("What's your name")
+let username
 let channelLink = document.getElementById("channel1")
 let channelNum;
+let user
 
-switch (channelLink.id){
+if (sessionStorage.getItem('user') === null) {
+    username = window.prompt("What's your name")
+
+} else {
+    username = JSON.parse(sessionStorage.getItem('user')).username
+}
+
+switch (channelLink.id) {
     case 'channel1':
         channelNum = 1
         break;
@@ -11,14 +19,52 @@ switch (channelLink.id){
         break;
 }
 
-let user = {
-    'username' : username,
-    'channelId' : channelNum
+user = {
+    'username': username,
+    'channelId': channelNum
 }
+sessionStorage.setItem('user', JSON.stringify(user))
 
-sessionStorage.clear()
-sessionStorage.setItem('user',JSON.stringify(user))
 
+/*
+fetch('/are-there-users-check')
+    .then(response => response.json())
+    .then(data => {
+        console.log(data)
+        if (data) {
+            sessionStorage.clear()
+        }
+    })
+    .catch(error => {
+        console.error('Error:', error);
+    });
+*/
+
+
+/*
+
+fetch('/does-username-exist', {
+    method: 'POST',
+    headers: {
+        'Content-Type': 'application/json'
+    },
+    body: sessionStorage.getItem('user')
+})
+    .then(response => response.json())
+    .then(data => {
+        if (data) {
+            console.log('true');
+        } else {
+            sessionStorage.clear()
+        }
+    })
+    .catch(error => {
+        console.error('Error:', error);
+    });
+*/
+
+
+channelLink.addEventListener('click', () => {
 
     fetch(`http://localhost:8080/add-user`, {
         method: "POST",
@@ -27,7 +73,11 @@ sessionStorage.setItem('user',JSON.stringify(user))
         },
         body: JSON.stringify(user)
     })
-       .then((response) => response.json())
+        .then((response) => response.json())
         .then((data) => {
             console.log(data)
-    })
+        })
+
+})
+
+
